fix: rename GraphQLExecutor to AbstractGraphQLExecutor (#149)
