+++ code/src/main/kotlin/com/expediagroup/sdk/lodgingconnectivity/sandbox/property/operation/UpdateSandboxPropertyOperation.kt
@@ -17,7 +17,7 @@
 package com.expediagroup.sdk.lodgingconnectivity.sandbox.property.operation
 
 import com.expediagroup.sdk.core.model.exception.service.ExpediaGroupServiceException
-import com.expediagroup.sdk.graphql.common.GraphQLExecutor
+import com.expediagroup.sdk.graphql.common.AbstractGraphQLExecutor
 import com.expediagroup.sdk.graphql.model.response.RawResponse
 import com.expediagroup.sdk.graphql.model.response.Response
 import com.expediagroup.sdk.lodgingconnectivity.sandbox.operation.SandboxUpdatePropertyMutation
@@ -41,16 +41,16 @@ data class UpdateSandboxPropertyResponse(
 /**
  * Executes [SandboxUpdatePropertyMutation] GraphQL mutation to modify the details of an existing sandbox property.
  *
- * This function uses the provided [GraphQLExecutor] to execute the mutation and returns an [UpdateSandboxPropertyResponse]
+ * This function uses the provided [AbstractGraphQLExecutor] to execute the mutation and returns an [UpdateSandboxPropertyResponse]
  * containing both the updated sandbox property data and the full raw response.
  *
- * @param graphQLExecutor The [GraphQLExecutor] responsible for executing the GraphQL mutation.
+ * @param graphQLExecutor The [AbstractGraphQLExecutor] responsible for executing the GraphQL mutation.
  * @param input The [UpdatePropertyInput] containing the details of the property update.
  * @return An [UpdateSandboxPropertyResponse] containing the updated sandbox property data and the full raw response.
  * @throws [ExpediaGroupServiceException] If an error occurs during the mutation execution.
  */
 fun updateSandboxPropertyOperation(
-    graphQLExecutor: GraphQLExecutor,
+    graphQLExecutor: AbstractGraphQLExecutor,
     input: UpdatePropertyInput
 ): UpdateSandboxPropertyResponse {
     val operation = SandboxUpdatePropertyMutation(input)
