+++ code/src/main/kotlin/com/expediagroup/sdk/lodgingconnectivity/sandbox/reservation/operation/UpdateSandboxReservationOperation.kt
@@ -17,7 +17,7 @@
 package com.expediagroup.sdk.lodgingconnectivity.sandbox.reservation.operation
 
 import com.expediagroup.sdk.core.model.exception.service.ExpediaGroupServiceException
-import com.expediagroup.sdk.graphql.common.GraphQLExecutor
+import com.expediagroup.sdk.graphql.common.AbstractGraphQLExecutor
 import com.expediagroup.sdk.graphql.model.response.RawResponse
 import com.expediagroup.sdk.graphql.model.response.Response
 import com.expediagroup.sdk.lodgingconnectivity.sandbox.operation.SandboxUpdateReservationMutation
@@ -40,16 +40,16 @@ data class UpdateSandboxReservationResponse(
 /**
  * Executes [SandboxUpdateReservationMutation] GraphQL mutation to modify the details of an existing sandbox reservation.
  *
- * This function uses the provided [GraphQLExecutor] to execute the mutation and returns an [UpdateSandboxReservationResponse]
+ * This function uses the provided [AbstractGraphQLExecutor] to execute the mutation and returns an [UpdateSandboxReservationResponse]
  * containing both the updated reservation data and the full raw response.
  *
- * @param graphQLExecutor The [GraphQLExecutor] responsible for executing the GraphQL mutation.
+ * @param graphQLExecutor The [AbstractGraphQLExecutor] responsible for executing the GraphQL mutation.
  * @param input The [UpdateReservationInput] containing the new details for the reservation.
  * @return An [UpdateSandboxReservationResponse] containing the updated reservation data and the full raw response.
  * @throws [ExpediaGroupServiceException] If an error occurs during the mutation execution.
  */
 fun updateSandboxReservationOperation(
-    graphQLExecutor: GraphQLExecutor,
+    graphQLExecutor: AbstractGraphQLExecutor,
     input: UpdateReservationInput
 ): UpdateSandboxReservationResponse {
     val operation = SandboxUpdateReservationMutation(input)
