+++ code/src/main/kotlin/com/expediagroup/sdk/lodgingconnectivity/supply/reservation/ReservationClient.kt
@@ -17,9 +17,9 @@
 package com.expediagroup.sdk.lodgingconnectivity.supply.reservation
 
 import com.expediagroup.sdk.core.model.exception.service.ExpediaGroupServiceException
-import com.expediagroup.sdk.graphql.common.DefaultGraphQLExecutor
-import com.expediagroup.sdk.lodgingconnectivity.common.GraphQLClient
 import com.expediagroup.sdk.graphql.common.GraphQLExecutor
+import com.expediagroup.sdk.lodgingconnectivity.common.GraphQLClient
+import com.expediagroup.sdk.graphql.common.AbstractGraphQLExecutor
 import com.expediagroup.sdk.lodgingconnectivity.common.RequestExecutor
 import com.expediagroup.sdk.lodgingconnectivity.configuration.ClientConfiguration
 import com.expediagroup.sdk.lodgingconnectivity.configuration.ClientEnvironment
@@ -63,7 +63,7 @@ class ReservationClient(config: ClientConfiguration) : GraphQLClient() {
 
     override val apiEndpoint = EndpointProvider.getSupplyApiEndpoint(config.environment)
 
-    override val graphQLExecutor: GraphQLExecutor = DefaultGraphQLExecutor(
+    override val graphQLExecutor: AbstractGraphQLExecutor = GraphQLExecutor(
         requestExecutor = RequestExecutor(config, apiEndpoint),
         serverUrl = apiEndpoint.endpoint
     )
